+++ src/main/java/com/github/idea/ginkgo/GinkgoRunConfigurationProducer.java
@@ -25,6 +25,10 @@ public class GinkgoRunConfigurationProducer extends LazyRunConfigurationProducer
 
     @Override
     protected boolean setupConfigurationFromContext(@NotNull GinkgoRunConfiguration configuration, @NotNull ConfigurationContext context, @NotNull Ref<PsiElement> sourceElement) {
+        if (context.getPsiLocation() != null && context.getPsiLocation().getParent() != null) {
+            return false;
+        }
+
         if (!(context.getPsiLocation().getParent().getParent() instanceof GoCallExpr)) {
             return false;
         }